UPDATE launch files bringup and library
