+++ sobit_edu_library/launch/library_server.launch.py
@@ -3,12 +3,26 @@ from ament_index_python.packages import get_package_share_directory
 
 from launch import LaunchDescription
 from launch_ros.actions import Node
+from launch.actions import DeclareLaunchArgument, OpaqueFunction
+from launch.substitutions import LaunchConfiguration
 
 
 def generate_launch_description():
-    robot_name = "sobit_edu"
-    robot_id = 0
+    arg_robot_name = DeclareLaunchArgument('robot_name', default_value='sobit_edu')
+    arg_enable_gz = DeclareLaunchArgument('enable_gz', default_value='True')
 
+    return LaunchDescription([
+        arg_robot_name,
+        arg_enable_gz,
+        OpaqueFunction(function = launch_gz),
+    ])
+
+def launch_gz(context, *args, **kwargs):
+    robot_name = LaunchConfiguration('robot_name').perform(context)
+    enable_gz = LaunchConfiguration('enable_gz').perform(context)
+    robot_id = 0
+    remappings = [('commands/velocity', 'diff_controller/cmd_vel_unstamped'),
+                 ('odom', 'diff_controller/odom')]
     pose_config = os.path.join(
         get_package_share_directory("sobit_edu_library"),
         "config",
@@ -20,7 +34,9 @@ def generate_launch_description():
         executable="joint_action_server",
         name="joint_action_server",
         namespace=robot_name if robot_id == 0 else f"{robot_name}_{robot_id}",
-        parameters=[pose_config],
+        parameters=[pose_config,
+            {"use_sim_time": True if enable_gz == 'True' else False},
+        ],
         output="screen",
     )
 
@@ -29,11 +45,15 @@ def generate_launch_description():
         executable="wheel_action_server",
         name="wheel_action_server",
         namespace=robot_name if robot_id == 0 else f"{robot_name}_{robot_id}",
+        parameters=[
+            {"use_sim_time": True if enable_gz == 'True' else False},
+        ],
+        remappings=remappings if enable_gz == 'True' else [],
         output="screen",
     )
 
 
-    return LaunchDescription([
+    return [
         joint_action_server_node,
         wheel_action_server_node,
-    ])
+    ]
